SL: - removed read time delay - refactored OffsetFetcher to GlobalLatestOffsetCache - updated tests to reflect changes
